feat(icon): added svg icon picker
