+++ CHANGELOG.zh_CN.md
@@ -1,5 +1,9 @@
 ## Wip
 
+### ✨ Features
+
+- 图标选择器新增 svg 模式
+
 ### ✨ Refactor
 
 - 重构项目以解决循环依赖项导致的热更新问题
