+++ package.json
@@ -53,7 +53,7 @@
   "devDependencies": {
     "@commitlint/cli": "^12.0.1",
     "@commitlint/config-conventional": "^12.0.1",
-    "@iconify/json": "^1.1.313",
+    "@iconify/json": "^1.1.314",
     "@ls-lint/ls-lint": "^1.9.2",
     "@purge-icons/generated": "^0.7.0",
     "@types/crypto-js": "^4.0.1",
@@ -110,7 +110,7 @@
     "vite-plugin-purge-icons": "^0.7.0",
     "vite-plugin-pwa": "^0.5.6",
     "vite-plugin-style-import": "^0.8.1",
-    "vite-plugin-svg-icons": "^0.3.3",
+    "vite-plugin-svg-icons": "^0.3.4",
     "vite-plugin-theme": "^0.4.8",
     "vite-plugin-windicss": "0.7.1",
     "vue-eslint-parser": "^7.6.0",
@@ -120,7 +120,7 @@
     "//": "Used to install imagemin dependencies, because imagemin may not be installed in China.If it is abroad, you can delete it",
     "bin-wrapper": "npm:bin-wrapper-china",
     "esbuild": "0.8.57",
-    "rollup": "2.40.0"
+    "rollup": "2.41.0"
   },
   "repository": {
     "type": "git",
