+++ src/design/index.less
@@ -37,6 +37,7 @@ a:focus,
 a:active,
 button,
 div,
+svg,
 span {
   outline: none !important;
 }
