+++ src/views/demo/feat/icon/index.vue
@@ -21,15 +21,21 @@
       </div>
     </CollapseContainer>
 
-    <CollapseContainer title="svg Sprite" class="my-5">
+    <CollapseContainer title="svg 雪碧图" class="my-5">
       <div class="flex justify-around flex-wrap">
         <SvgIcon name="test" size="32" />
       </div>
     </CollapseContainer>
 
-    <CollapseContainer title="图标选择器" class="my-5">
+    <CollapseContainer title="图标选择器(Iconify)" class="my-5">
       <div class="flex justify-around flex-wrap">
         <IconPicker />
+      </div>
+    </CollapseContainer>
+
+    <CollapseContainer title="图标选择器(Svg)" class="my-5">
+      <div class="flex justify-around flex-wrap">
+        <IconPicker mode="svg" />
       </div>
     </CollapseContainer>
